Ajax CRUD

+++ app/Http/Controllers/PostController.php
@@ -2,17 +2,28 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\VideoViewer;
+use App\Http\Requests\PostRequest;
 use App\Models\Post;
+use App\Models\User;
+use App\Models\Video;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
         $posts = Post::get();
         return view('posts', compact('posts'));
+
     }
 
     public function create()
@@ -50,9 +61,10 @@ class PostController extends Controller
         $post->title = $request->title;
         $post->desc = $request->desc;
         $post->content = $request->content;
+        $post->user_id = auth()->user()->id;
         $post->image = $imageName;
         $post->save();
-        return redirect('posts')->with('success', 'Post Created');
+        return redirect(route('index.home'))->with('success', 'Post Created');
     }
 
     public function show($id)
@@ -64,24 +76,38 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::where('id', $id)->first();
+
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect(route('index.home'))->with('error', 'Unauthorized Page');
+        }
+
         return view('edit', compact('post'));
     }
 
-    public function update(Request $request, $id)
+    public function update(PostRequest $request, $id)
     {
-        $data = Validator::make($request->all(),[
-            'title' => 'required|string|max:100',
-            'desc' => 'required|string',
-            'content' => 'required|string',
-            'image' => 'nullable|image|mimes:png,jpg,jpeg'
-        ]);
-        //Lw el data feha errors
-        if ($data->fails())
-        {
-            return back()
-            ->withErrors($data)
-            ->withInput();
-        }
+        // $rules = [
+        //     'title' => 'required|string|max:100',
+        //     'desc' => 'required|string',
+        //     'content' => 'required|string',
+        //     'image' => 'nullable|image|mimes:png,jpg,jpeg'
+        // ];
+
+        // $data = Validator::make($request->all(), $rules,[
+        //     'desc.required' => 'The description field is required.'
+        // ]);
+
+        // //Lw el data feha errors
+        // if ($data->fails())
+        // {
+        //     // return $data->errors()->first();
+        //     return back()
+        //     ->withErrors($data)
+        //     ->withInput();
+        // }
+
         $post = Post::where('id', $id)->first();
         $old_name = $post->image;
         if ($request->hasFile('image'))
@@ -101,20 +127,34 @@ class PostController extends Controller
         $post->title = $request->title;
         $post->desc = $request->desc;
         $post->content = $request->content;
+        $post->user_id = auth()->user()->id;
         $post->image = $imageName;
         $post->save();
-        return redirect('posts')->with('success', 'Post Updated');
+        return redirect(route('index.home'))->with('success', 'Post Updated');
     }
 
     public function destroy($id)
     {
         $post = Post::where('id', $id)->first();
 
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect(route('index.home'))->with('error', 'Unauthorized Page');
+        }
+
         //Delete image
         $old_name = $post->image;
         Storage::disk('uploads')->delete($old_name);
 
         $post->delete();
-        return redirect('posts')->with('success', 'Post Removed');
+        return redirect(route('index.home'))->with('success', 'Post Removed');
+    }
+
+    public function getVideo()
+    {
+        $video = Video::first();
+        event(new VideoViewer($video));
+        return view ('video')->with('video', $video);
     }
 }
